Replace exercise 05 with 05-extra-1
The current version of Exercise 5 was the code from final/02.extra-4.tsx
