+++ src/final/05.tsx
@@ -1,94 +1,42 @@
 // DOM Side-Effects
-// http://localhost:3000/isolated/final/02.extra-4.tsx
+// 💯 Side-effect cleanup
+// http://localhost:3000/isolated/final/05.tsx
 
 import * as React from 'react'
-import {useLocalStorageState} from '../utils'
+import VanillaTilt from 'vanilla-tilt'
 
-function UsernameForm({
-  initialUsername = '',
-  onSubmitUsername,
-}: {
-  initialUsername?: string
-  onSubmitUsername: (username: string) => void
-}) {
-  const [username, setUsername] = useLocalStorageState(
-    'username',
-    initialUsername,
-  )
-  const [touched, setTouched] = React.useState(false)
+interface HTMLVanillaTiltElement extends HTMLDivElement {
+  vanillaTilt: VanillaTilt
+}
 
-  const usernameIsLowerCase = username === username.toLowerCase()
-  const usernameIsLongEnough = username.length >= 3
-  const usernameIsShortEnough = username.length <= 10
-  const formIsValid =
-    usernameIsShortEnough && usernameIsLongEnough && usernameIsLowerCase
-
-  const displayErrorMessage = touched && !formIsValid
-
-  const usernameInputRef = React.useRef<HTMLInputElement>(null)
+function Tilt({children}: {children: React.ReactNode}) {
+  const tiltRef = React.useRef<HTMLVanillaTiltElement>(null)
 
   React.useEffect(() => {
-    if (displayErrorMessage) usernameInputRef.current?.focus()
-  }, [displayErrorMessage])
-
-  let errorMessage = null
-  if (!usernameIsLowerCase) {
-    errorMessage = 'Username must be lower case'
-  } else if (!usernameIsLongEnough) {
-    errorMessage = 'Username must be at least 3 characters long'
-  } else if (!usernameIsShortEnough) {
-    errorMessage = 'Username must be no longer than 10 characters'
-  }
-
-  function handleSubmit(event: React.FormEvent<HTMLFormElement>) {
-    event.preventDefault()
-    setTouched(true)
-    if (!formIsValid) return
-
-    onSubmitUsername(username)
-  }
-
-  function handleChange(event: React.ChangeEvent<HTMLInputElement>) {
-    setUsername(event.currentTarget.value)
-  }
-
-  function handleBlur() {
-    setTouched(true)
-  }
+    const {current: tiltNode} = tiltRef
+    if (tiltNode === null) return
+    const vanillaTiltOptions = {
+      max: 25,
+      speed: 400,
+      glare: true,
+      'max-glare': 0.5,
+    }
+    VanillaTilt.init(tiltNode, vanillaTiltOptions)
+    return () => tiltNode.vanillaTilt.destroy()
+  }, [])
 
   return (
-    <form name="usernameForm" onSubmit={handleSubmit} noValidate>
-      <div>
-        <label htmlFor="usernameInput">Username:</label>
-        <input
-          ref={usernameInputRef}
-          id="usernameInput"
-          type="text"
-          value={username}
-          onChange={handleChange}
-          onBlur={handleBlur}
-          pattern="[a-z]{3,10}"
-          required
-          aria-describedby={displayErrorMessage ? 'error-message' : undefined}
-        />
-      </div>
-      {displayErrorMessage ? (
-        <div role="alert" id="error-message">
-          {errorMessage}
-        </div>
-      ) : null}
-      <button type="submit">Submit</button>
-    </form>
+    <div ref={tiltRef} className="tilt-root">
+      <div className="tilt-child">{children}</div>
+    </div>
   )
 }
 
 function App() {
-  const onSubmitUsername = (username: string) =>
-    alert(`You entered: ${username}`)
   return (
-    <div style={{width: 400}}>
-      <UsernameForm onSubmitUsername={onSubmitUsername} />
-    </div>
+    <Tilt>
+      <div className="totally-centered">vanilla-tilt.js</div>
+    </Tilt>
   )
 }
